refactor contracts to program as they can contain transactions and scripts
